final changes for pmax 4.4-Lite II release

+++ sys/pmax/dev/scc.c
@@ -7,7 +7,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)scc.c	8.2 (Berkeley) 11/30/93
+ *	@(#)scc.c	8.3 (Berkeley) 06/02/95
  */
 
 /* 
@@ -394,7 +394,7 @@ sccwrite(dev, uio, flag)
 /*ARGSUSED*/
 sccioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
